Generate a MultiAssayExperiment object from multiple single-omics data

Generate a MultiAssayExperiment object from single-omics data matrix and metadata. Currently supports transcriptomics and proteomics data only.

generate_multiassay(
  rawdata_rna,
  rawdata_protein,
  individual_to_sample = FALSE,
  map_rna,
  map_protein,
  metadata_rna,
  metadata_protein,
  individual_metadata,
  map_by_column,
  rna_qc_data = FALSE,
  rna_qc_data_matrix,
  organism = "human"
)

Arguments

rawdata_rna

A data-frame containing raw RNA count where rows and columns correspond to genes and samples respectively

rawdata_protein

A data-frame containing raw protein abundance where rows and columns correspond to genes and samples respectively

individual_to_sample

Logical whether individual ID and sample names in raw data matrices are the same. If they are different, mpa_rna and map_protein dataframes should be provided. Default to FALSE.

map_rna

A data-frame of two columns named primary and colname where primary should contain unique ID name with a link to individual metadata and colname is the column names of the rawdata_rna data-frame

map_protein

A data-frame of two columns named primary and colname where primary should contain unique ID name with a link to individual metadata and colname is the column names of the rawdata_protein data-frame

metadata_rna

A data-frame containing rna assay specific metadata where rownames are same as the colnames of rawdata_rna data.frame

metadata_protein

A data-frame containing protein assay specific metadata where rownames are same as the colnames of rawdata_protein data.frame

individual_metadata

A data-frame containing sample level metadata

map_by_column

The common column name to link metadata_rna and metadata_protein to the individual_metadata

rna_qc_data

Logical whether to add rna QC data.

rna_qc_data_matrix

A data.frame containing RNA qc level data where rownames are same as the colnames of rawdata_rna data.frame.

organism

The organism type for transcripts and proteins. Possible values are human and mouse. Default to human.

Value

a MultiAssayExperiment object

See also