Functional enrichment analysis using enrichR

Performs impacted pathway analysis with a list of genes.

pathway_analysis_enrichr(
  interest_gene = NULL,
  project_name = NULL,
  enrichment_database = c("GO_Molecular_Function_2021", "GO_Cellular_Component_2021",
    "GO_Biological_Process_2021", "WikiPathways_2021_Human", "Reactome_2016",
    "KEGG_2021_Human", "MSigDB_Hallmark_2020", "BioCarta_2016"),
  is_output = FALSE,
  output_dir = ".",
  min_overlap = 3,
  plot_n = 10
)

Arguments

interest_gene

vector containing genes of interest Column names should be gene, logFC, pval and padj respectively.

project_name

The name of the project to save.

enrichment_database

Name of the database for enrichment. User can specify one or more database names from enrichR::listEnrichrDbs().

is_output

If TRUE a folder will be created and results of enrichment analysis will be saved otherwise a R list will be returned. Default FALSE

output_dir

Path for the output directory. Default is current dir.

min_overlap

minimum overlap between genes and genesets to keep.

plot_n

number of genesets to plot

Value

enrichment_result a list of data.frames containing enrichment output and a list of plots of top 10 significant genesets.

See also