Process bulk proteomics data

Process bulk proteomics data

process_protein(
  multiassay,
  filter = TRUE,
  min_sample = 0.5,
  dependent = "diagnosis",
  levels = c("Control", "AD"),
  imputation = "minimum_value",
  remove_feature_outliers = TRUE,
  batch_correction = TRUE,
  batch = "batch",
  correction_method = c("Combat", "Median_centering"),
  remove_sample_outliers = TRUE,
  denoise = FALSE,
  covariates = c("sex", "age")
)

Arguments

multiassay

Multiassay experiment object generated by Omix generate_multiassayfunction

filter

Logical whether to perform protein filtering

min_sample

Percentage of samples that must non missing protein value, or the protein is filtered out protein value, or else the protein is excluded

dependent

Dependent variable for groups. If no specific dependent variable exist, set parameter as NULL

levels

Character vector with reference group as first element. Set parameter as NULL if dependent is NULL.

imputation

Possible values are distribution, minimum_value,zero default to minimum_value.

remove_feature_outliers

Logical whether to remove feature outliers

batch_correction

Logical whether to perform technical batch correction

batch

Technical batch

correction_method

Correction methods available to counteract batch effects. Possible values are Combat or median_centering

remove_sample_outliers

Logical whether to remove sample outliers

denoise

Logical whether to denoise biological covariates

covariates

Covariates used in linear mixed model for denoising

Value

a MultiAssayExperiment object with protein_processed slot

See also