Process batch transcriptomics data

Process batch transcriptomics data

process_rna(
  multiassay,
  transformation = c("vst", "rlog", "log2"),
  protein_coding = TRUE,
  filter = TRUE,
  min_count = 10,
  min_sample = 0.5,
  dependent = "diagnosis",
  levels = NULL,
  covariates,
  batch_correction = TRUE,
  batch = "batch",
  remove_sample_outliers = FALSE
)

Arguments

multiassay

Multiassay experiment object generated by Omix generate_multiassayfunction

transformation

transformation method to choose from. Available methods are vst, rlog, and log2.

protein_coding

Logical whether to filter out non protein coding genes

filter

Logical whether to perform gene filtering

min_count

Minimum count

min_sample

Percentage of samples that have at least the min_count, or else the gene is filtered out

dependent

Dependent variable for the DeSEQ2 analysis, usually the disease group variable

levels

Character vector with reference group as first element. Set parameter as NULL if dependent is NULL.

covariates

Technical and biological covariates to use for DESEQ2 analysis and denoising

batch_correction

Logical whether to perform batch correction. The processed matrix returned will be denoised for designated covariates and technical batch.

batch

Technical batch, use batch2 for another technical batch

remove_sample_outliers

Logical whether to remove sample outliers

Value

a MultiAssayExperiment object with rna_processed slot

See also