Compare differential expressed genes and proteins

Compare differential expressed genes and proteins

single_omic_comparisons(
  multiassay,
  threshold = 0.05,
  pvalue = c("padj", "pvalue", "all"),
  filtering_options = "both_significant",
  additional_database_gmt = NULL
)

Arguments

multiassay

Multiassay experiment object generated by Omix

threshold

is the pvalue threshold. Default to 0.05

pvalue

Comparison reference.can be set to pvalue or padj, to consider the normal p values or the ones after multiple testing correction, respectively. With pvalue == all, no p value threshold is applied and all genes will be compared. all genes are kept and only the direction is compared.

filtering_options

default to NULL. if set to both_significantthe returned dataframe will only display genes that are significant at both transcriptomics and proteomics levels. Setting the parameter to eitherwill return genes that are significant in at least one layer.

additional_database_gmt

provide GMT file path for integrated pathway enrichment

Value

A list object with dataframe and plot slots

See also