Patient specific vertical integration

Patient specific vertical integration

vertical_integration(
  multiassay,
  slots = c("rna_processed", "protein_processed"),
  integration,
  intersect_genes = FALSE,
  ID_type = "gene_name",
  dependent = "Group",
  levels = c("Control", "Case"),
  design = "cor",
  ncomp = 2,
  range = list(mRNA = seq(5, 100, by = 10), proteins = seq(5, 100, by = 10)),
  list.keepX = list(mRNA = c(50), proteins = c(50)),
  num_factors = 10,
  time = "pseudotime",
  scale_views = TRUE,
  try.N.clust = 2:4,
  most_variable_feature = FALSE
)

Arguments

multiassay

Multiassay experiment object generated by Omix

slots

Default are processed slots

integration

Possible integration methods are MOFA, DIABLO,sMBPLS,iCluster,MEIFESTO

intersect_genes

Logical whether to keep intersecting features for integration

ID_type

Default to gene_name

dependent

Dependent variable for the DeSEQ2 analysis, usually the disease group variable

levels

Character vector with reference group as first element. Set parameter as NULL if dependent is NULL.

design

Available options "cor" and "full". Default is "cor". Design matrix (design = "full"): The strength of all relationships between dataframes is maximised (= 1) – a “fully connected” design. If design is set on cor, the correlation between PC1 of each dataset will be set for the design matrix

ncomp

Number of components in DIABLO,sMBPLS

range

List of the range of numbers of features to keep in the tuning phase. First element must be for rna, second for proteins

list.keepX

if integration == sMBPLS, number of features to keep per view.

num_factors

if integration == MOFA, number of factors. Default to 10.

time

Pseudotime covariate

scale_views

Logical whether to scale omic views

try.N.clust

number of cluster to tune in iCluster

most_variable_feature

Logical whether to keep most variable features

Value

Returns an integrated object in multiassay@metadata$integration to be used for further analysis

See also