Get multi-omics comprehensive heatmap

This function vertically concatenates multiple heatmap derived from each omics data. getMoHeatmap supports customized column annotation and is able to mark the selected features if indicated.

multiomics_clustering_heatmap(
  data = NULL,
  is.binary = c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE),
  row.title = c("Data1", "Data2", "Data3", "Data4", "Data5", "Data6"),
  legend.name = c("Data1", "Data2", "Data3", "Data4", "Data5", "Data6"),
  clust.res = NULL,
  clust.dend = NULL,
  show.col.dend = TRUE,
  show.colnames = FALSE,
  show.row.dend = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  show.rownames = c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE),
  clust.dist.row = c("pearson", "pearson", "pearson", "pearson", "pearson", "pearson"),
  clust.method.row = c("ward.D", "ward.D", "ward.D", "ward.D", "ward.D", "ward.D"),
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  color = rep(list(c("#00FF00", "#000000", "#FF0000")), length(data)),
  annCol = NULL,
  annColors = NULL,
  annRow = NULL,
  width = 6,
  height = 4,
  fig.path = getwd(),
  fig.name = "moheatmap"
)

Arguments

data

A list of data frame or matrix storing multiple omics data with rows for features and columns for samples.

is.binary

A logicial vector to indicate if the subdata is binary matrix of 0 and 1 such as mutation.

row.title

A string vector to assign titles for each subdata.

legend.name

A string vector to assign legend title for each subdata.

clust.res

A clust.res object returned by getMOIC() with one specified algorithm or get\%algorithm_name\% or getConsensusMOIC() with a list of multiple algorithms.

clust.dend

A dendrogram object returned returned by getMOIC() with one specified algorithm or get\%algorithm_name\% or getConsensusMOIC() with a list of multiple algorithms.

show.col.dend

A logical vector to indicate if showing the dendrogram for column at the top of heatmap.

show.colnames

A logical vector to indicate if showing the names for column at the bottom of heatmap.

show.row.dend

A logical vector to indicate if showing the dendrogram for row of each subdata.

show.rownames

A logical vector to indicate if showing the names for row of each subdata.

clust.dist.row

A string vector to assign distance method for clustering each subdata at feature dimension.

clust.method.row

A string vector to assign clustering method for clustering each subdata at feature dimension.

clust.col

A string vector storing colors for annotating each subtype at the top of heatmap.

color

A list of string vectors storing colors for each subheatmap of subdata.

annCol

A data.frame storing annotation information for samples with exact the same sample order with data parameter.

annColors

A list of string vectors for colors matched with annCol.

annRow

A list of string vectors to indicate which features belong to which subdata should be annotated specifically in subheatmap.

width

An integer value to indicate the width for each subheatmap with unit of cm.

height

An integer value to indicate the height for each subheatmap with unit of cm.

fig.path

A string value to indicate the output path for storing the comprehensive heatmap.

fig.name

A string value to indicate the name of the comprehensive heatmap.

Value

A pdf of multi-omics comprehensive heatmap

References

Gu Z, Eils R, Schlesner M (2016). Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.

See also