Patient specific vertical integration

Patient specific vertical integration

Usage

vertical_integration(
  multiassay,
  slots = c("rna_processed", "protein_processed"),
  integration = c("MOFA", "DIABLO", "sMBPLS", "iCluster", "MEIFESTO"),
  intersect_genes = FALSE,
  ID_type = "gene_name",
  dependent = "Group",
  levels = c("Control", "Case"),
  design = c("cor", "full"),
  ncomp = 2,
  range = list(mRNA = seq(5, 100, by = 10), proteins = seq(5, 100, by = 10)),
  list.keepX = list(mRNA = c(50), proteins = c(50)),
  num_factors = 10,
  time = "pseudotime",
  scale_views = TRUE,
  try.N.clust = 2:4,
  most_variable_feature = FALSE,
  ...
)

Arguments

multiassay

Multiassay experiment object generated by Omix

slots

Default are processed slots

integration

Possible integration methods are MOFA, DIABLO,sMBPLS,iCluster,MEIFESTO

intersect_genes

Logical whether to keep intersecting features for integration

design

Design matrix (design = "full"): The strength of all relationships between dataframes is maximised (= 1) – a “fully connected” design. If design is set on cor, the correlation between PC1 of each dataset will be set for the design matrix

ncomp

Number of components in DIABLO,sMBPLS

range

List of the range of numbers of features to keep in the tuning phase. First element must be for rna, second for proteins

try.N.clust

number of cluster to tune in iCluster

Value

Returns an integrated object in multiassay@metadata$integration to be used for further analysis